use FromRemoteRepoMixin for EnrichmentManager
