+++ gto/api.py
@@ -518,18 +518,19 @@ def _show_versions(  # pylint: disable=too-many-locals
     return versions_, "keys"
 
 
-@clone_on_remote_repo
 def describe(
     repo: Union[str, Repo], name: str, rev: str = None
 ) -> List[EnrichmentInfo]:
     """Find enrichments for the artifact"""
     ref_type, parsed = parse_name_reference(name)
     if ref_type == NAME_REFERENCE.NAME:
-        return EnrichmentManager.from_repo(repo).describe(name=name, rev=rev)
+        with EnrichmentManager.from_repo(repo) as em:
+            return em.describe(name=name, rev=rev)
     if ref_type == NAME_REFERENCE.TAG:
         if rev:
             raise WrongArgs("Should not specify revision if you pass git tag")
-        return EnrichmentManager.from_repo(repo).describe(name=parsed[NAME], rev=name)
+        with EnrichmentManager.from_repo(repo) as em:
+            return em.describe(name=parsed[NAME], rev=name)
     raise NotImplementedError
