+++ gto/registry.py
@@ -1,7 +1,5 @@
 import logging
 import os
-from abc import abstractmethod
-from contextlib import contextmanager
 from pathlib import Path
 from typing import Optional, TypeVar, Union
 
@@ -33,7 +31,7 @@ from gto.exceptions import (
     VersionExistsForCommit,
     WrongArgs,
 )
-from gto.git_utils import cloned_git_repo, git_push_tag, is_url_of_remote_repo
+from gto.git_utils import FromRemoteRepoMixin, git_push_tag
 from gto.index import EnrichmentManager
 from gto.tag import (
     TagArtifactManager,
@@ -48,29 +46,6 @@ from gto.versions import SemVer
 TBaseEvent = TypeVar("TBaseEvent", bound=BaseEvent)
 
 
-class FromRemoteRepoMixin:
-    @classmethod
-    @abstractmethod
-    def _from_repo(cls, repo=Union[str, Repo], config: RegistryConfig = None):
-        pass
-
-    @classmethod
-    @contextmanager
-    def from_repo(cls, repo=Union[str, Repo], config: RegistryConfig = None):
-        if isinstance(repo, str) and is_url_of_remote_repo(repo=repo):
-            try:
-                with cloned_git_repo(repo=repo) as tmp_dir:
-                    yield cls._from_repo(repo=tmp_dir, config=config)
-            except (NotADirectoryError, PermissionError) as e:
-                raise e.__class__(
-                    "Are you using windows with python < 3.9? "
-                    "This may be the reason of this error: https://bugs.python.org/issue42796. "
-                    "Consider upgrading python."
-                ) from e
-        else:
-            yield cls._from_repo(repo=repo, config=config)
-
-
 class GitRegistry(BaseModel, FromRemoteRepoMixin):
     repo: git.Repo
     artifact_manager: TagArtifactManager
@@ -83,7 +58,7 @@ class GitRegistry(BaseModel, FromRemoteRepoMixin):
         arbitrary_types_allowed = True
 
     @classmethod
-    def _from_repo(cls, repo=Union[str, Repo], config: RegistryConfig = None):
+    def _from_repo(cls, repo: Union[str, Repo], config: RegistryConfig = None):
         if isinstance(repo, str):
             try:
                 repo = git.Repo(repo, search_parent_directories=True)
